Revert "Feat/aggregate cron"
