+++ server/utils/queue.js
@@ -1,11 +1,9 @@
-import { aggregateCheck } from '@server/cronJobs/aggregateJob';
 import Bull from 'bull';
 import moment from 'moment';
 const queues = {};
 
 export const QUEUE_NAMES = {
-  SCHEDULED_JOB: 'scheduledJob',
-  AGGREGATE_CHECK: 'aggregateCheck'
+  SCHEDULED_JOB: 'scheduledJob'
 };
 
 export const QUEUE_PROCESSORS = {
@@ -14,16 +12,7 @@ export const QUEUE_PROCESSORS = {
       message: job.data.message
     });
     done();
-  },
-  [QUEUE_NAMES.AGGREGATE_CHECK]: (job, done) => {
-    console.log('Aggregate job is getting executed');
-    aggregateCheck();
-    done();
   }
-};
-
-const CRON_EXPRESSIONS = {
-  MIDNIGHT: '0 0 * * *'
 };
 
 export const initQueues = () => {
@@ -32,7 +21,6 @@ export const initQueues = () => {
     queues[queueName] = getQueue(queueName);
     queues[queueName].process(QUEUE_PROCESSORS[queueName]);
   });
-  queues[QUEUE_NAMES.AGGREGATE_CHECK].add({}, { repeat: { cron: CRON_EXPRESSIONS.MIDNIGHT } });
 };
 export const getQueue = queueName => {
   if (!queues[queueName]) {
